LayoutTests/platform/ios/fast/text/international/system-language/navigator-language/navigator-language-hi-expected.txt

FAIL navigator.language should be hi-in. Was hi.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
